# lib/api-zod/src/generated/types/changePasswordBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JNMT Student Hub API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ChangePasswordBody {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}