# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JNMT Student Hub API specification
 * OpenAPI spec version: 0.2.0
 */

export interface RegisterBody {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}